Refactor types with an index file to make them better organized and clean up imports
