+++ src/player-ui/components/ControlButtons/index.tsx
@@ -5,7 +5,7 @@ import styled, { css, useTheme } from 'styled-components';
 import { useExpanded } from '~player-ui/contexts/Expanded';
 import type { Expandable } from '~player-ui/types';
 import { expandedStyle } from '~player-ui/util/expandedStyle';
-import { RepeatMode } from '~types/RepeatMode';
+import { RepeatMode } from '~types';
 
 interface ControlButtonsProps {
   isPlaying: boolean;
