+++ src/types/Events.ts
@@ -1,4 +1,4 @@
-import type { PlayerState, SongInfo } from './PlayerState';
+import type { PlayerState, SongInfo } from '.';
 
 export enum EventMessage {
   SONG_INFO_UPDATED = 'SONG_INFO_UPDATED',
