+++ src/types/QueueItem.ts
@@ -1,0 +1,6 @@
+import type { SongInfo } from '.';
+
+export interface QueueItem {
+  isPlaying: boolean;
+  songInfo: SongInfo;
+}
