+++ src/types/SessionDetails.ts
@@ -1,5 +1,4 @@
-import type { Listener } from './Listener';
-import type { RepeatMode } from './RepeatMode';
+import type { Listener, RepeatMode } from '.';
 
 export interface SessionDetails {
   hostId: string;
